Move deprecated Cloudflare v1
